reportcomponentadded
